requirement.txt
djangorestframework-simplejwt
djangorestframework

drf-yasg
